[raw data][airflow] Add raw data file chunking entrypoint (Recidiviz/recidiviz-data#30884)
## Description of the change

Closing https://github.com/Recidiviz/recidiviz-data/pull/30528 in favor
of this pr.

Moving to kubernetes to have better control over the resources this task
uses

To address previous comments regarding `PreImportNormalizationType` -
only files that have a `PreImportNormalizationType` should be passed to
this step so the `normalization_type` should never be `None`., but the
function
`PreImportNormalizationType.required_pre_import_normalization_type` can
return `None` so the field was made optional. I've added
normalization_type to `RequiresNormalizationFile` type (the input to the
file chunking task) so we can remove this call.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

https://github.com/Recidiviz/recidiviz-data/issues/29946

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7d2ab89f51b7912e0e82465061b8361ead352b39