[JII] Deploy the prototypes server to Cloud Run. (Recidiviz/recidiviz-data#31054)
## Description of the change

Deploy the prototypes server to Cloud Run.

Add the cloud run trigger yaml for deploying the prototypes server to
cloud run, and the python script to execute the cloud build trigger.

Zenhub TIcket Recidiviz/recidiviz-data#30939

See the prototypes-server Cloud Trigger
[here](https://console.cloud.google.com/cloud-build/triggers;region=global/edit/fb53f433-8f13-48be-b249-89dd6c8eb1ba?project=recidiviz-staging).
The config file is currently inlined into the trigger, but once this PR
is submitted we will replace the inlined config with a reference to this
Recidiviz/recidiviz-data yaml file.

## Testing

Successfully deployed the Prototypes server to recidiviz-staging.
The hello world endpoint can be visited at
https://prototypes-qqec6jbn6a-uc.a.run.app/

Note: We are skipping IAM permissions for this endpoint and will use an
OAuth wall instead.
GitOrigin-RevId: 1d05d35af0909c07dcac4e0230e03d3347dbf671