[Workflows] Convert opp config fields from JSONB to JSON (Recidiviz/recidiviz-data#31076)
## Description of the change

Converts JSONB fields in the opportunity configuration into JSON. Denial
reasons and criteria copy are stored as dictionaries/objects with the
keys in meaningful order, but JSONB doesn't retain key ordering. JSON
does, at the expense of performance if we were actually querying on the
JSON values. We're not doing that, so this shouldn't have any negative
impact.

Another approach would be to store these fields as arrays instead, but
that would require reworking the schemas in a bunch of different places.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 455d61eb81930f7dc978930de9c8110a6ef87f51