[Justice Counts] Don't reference filename parameter in bytes object during Automated Bulk Upload (Recidiviz/recidiviz-data#31056)
The Bulk Upload for Automated Bulk Upload was encountering an error when
the Excel file was downloaded as a bytes object. The bytes object does
not have a filename parameter that was expected in
`SpreadsheetInterface.ingest_spreadsheet`, causing the sheet not to be
ingested ([Sentry
Error](https://recidiviz-inc.sentry.io/issues/5557467728/?project=4504532096516096&query=is%3Aunresolved+issue.priority%3A%5Bhigh%2C+medium%5D&referrer=issue-stream&stream_index=0)).
This was caused because I removed the filename parameter in the last PR,
thinking that I did not need it.

**Fix:**

- Introduced a new `file_name` parameter for SpreadsheetInterface to
handle the file name separately.
- Updated all references from `filename` to `file_name` to ensure
consistency and because "file name" is two words and was previously
referenced as one.

**Testing:**

Tested by uploading to staging via bulk upload and confirmed that the
sheet was uploaded without difficulties.

## Related issues

Closes Recidiviz/recidiviz-data#31048

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 72c636ea4ca2dfa24d3631514610e9fa2eb61f94