Fixes ambiguous columns bug in custom metrics explore (Recidiviz/recidiviz-data#31030)
## Description of the change

Fixes a bug (replicate
[here](https://recidiviz.cloud.looker.com/explore/recidiviz-123/workflows_impact_metrics?qid=yCl3VqiREUipyJAukdhvUf&origin=share-expanded&toggle=fil))
where there is an "ambiguous" column error if a unit of analysis index
column is selected in the workflows impact metrics explore. This error
was happening because index columns are not included in the join
condition, resulting in duplicate index columns being selected in cases
where there are multiple unit-of-observation metric cte's. To fix, I
only include unit of analysis columns (via the "column_mapping" field)
for the _first_ cte.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#31028

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: a607aec2d50ae97a6ac7b24d83986ae5c29e07d0