[Supervisor homepage] Pull officer avg caseload for the period into the querier (Recidiviz/recidiviz-data#30802)
An officer's avg_daily_population is the calculated avg caseload size on any given day in a given period. This value is included in every row in the officer aggregated metrics and pulled in as an individual metric in the supervision_officer_metrics view.

This PR pulls adds the current avg_daily_population as a top-level field on the officer entity. We do this by adding a join to the SOM table in the querier and selecting the most recent avg_daily_population value for an officer (aka the CURRENT value).

We'll be displaying this field on the officer's overview page and metric pages in the new supervisor homepage UI

GitOrigin-RevId: e444d7b61473b54ae33cecc199eb10bbc43aaaf6