[BigQuery] DML insert, take two (Recidiviz/recidiviz-data#31026)
## Description of the change

Un-reverts Recidiviz/recidiviz-data#30970. Incorrectly convinced myself that DML statements
could include a destination table or write disposition (the insert
clause specifies these) as it had worked using the emulator (oops).

a few additional notes:
- if the intersection between the two schemas does not include a
required field of the destination table that does not have a default
value, we'll see an error like this
`google.api_core.exceptions.BadRequest: 400 GET
https://bigquery.googleapis.com/bigquery/v2/projects/recidiviz-staging/queries/1bf62817-8ac6-4dac-a055-d342fa6d4731?maxResults=0&location=US&prettyPrint=false:
Required field file_id cannot be null`. If the column does have a
default value, it'll just fill it in with that.
- field type must match for all columns but because we always explicitly
specify the columns being inserted, the column ordering does not have to
match.
- removed `hydrate_missing_columns_with_null` as will be done
automatically.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Un-reverts Recidiviz/recidiviz-data#30970

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7dba1a4a4e10320ab26ce68dd51d0b1592ebf744