[Justice Counts] Access Agency Datapoints via the MetricSettingInterface class in all tests. (Recidiviz/recidiviz-data#30493)
## Description of the change

Access Agency Datapoints via the MetricSettingInterface class in all
tests.

Modify test suites to only access agency datapoints (or metric settings
entries) using the MetricSettingsInterface. This is a necessary step so
that our tests won't break when we start reading from the MetricSettings
table instead of the Agency Datapoint table.

It is also better practice to use the Interface methods instead of
directly writing the entries into the tables.

## Testing

All of the changes in this PR are a direct translation of testing logic
from using agency datapoints directly to using the equivalent
MetricSettingInterface method.

GitOrigin-RevId: 9845381e000fae40cf18a1e2d5ff5b33a500f02a