Creation of PSI Case Insights rates table (Recidiviz/recidiviz-data#30876)
## Description of the change

This PR writes a new table `sentencing.case_insights_rates` which will
serve as the basis for the numbers used to create the Case Insights
report. This table follows the following schema:
https://docs.google.com/document/d/1vexI5d_l_2ElbYpw-69jVJkK7B7VPbIeHUH_EIQ7olQ/edit.

This initial version does not have any roll-up logic. It has a row for
every individual fine-grained offense and risk bucket, and the
`recidivism_rollup` value is simply the `offense`. Roll-up logic will be
implemented in Recidiviz/recidiviz-data#30755 as a followup.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#30942

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a632998c1067093f9d65fb184591c01cf3bc1e24