[Insights] Join on email address for success_logins (Recidiviz/recidiviz-data#30844)
## Description of the change

QA caught a few users who logged into Insights but their login was not
being captured in this result set. The cause of this was that the user's
external_id in the Admin Panel was incorrect (those few are now fixed),
which means that the
`auth0_prod_action_logs.success_login.pseudonymized_id` was incorrect so
wasn't being joined to the supervisor correctly. This updates the join
to use email address instead. We're seeing an additional 164 logins in
this view now.

This has been run in the production jovergaag_usage sandbox.

Ticket to fix the incorrect ids is
[here](https://app.zenhub.com/workspaces/polaris-614c9f27fa19b10015ddd495/issues/gh/recidiviz/recidiviz-data/30843).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#30237

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 136cd0e845fa5cb99fd87633acc809b672ac7a47