[Dataflow] Fix `launch_env` checks when running inside pipeline steps (Recidiviz/recidiviz-data#31196)
## Description of the change

@brendali121 ran into an issue last week where she was unable to gate an
ingest view using the `launch_env` configuration.

The current method of detecting the project we're running in is based
off of the values of environment variables. These variables are not
guaranteed to be set in Dataflow workers as they process steps.

This change builds launch context from the pipeline parameters instead
of the environment variables.

I'm not fully sure on how we should unit test this change, it feels like
we'd need to determine if beam's `DirectRunner` also exhibits the same
issue.

This is a successful pipeline run of the view from Recidiviz/recidiviz-data#31025 that was
previously failing-
https://console.cloud.google.com/dataflow/jobs/us-west1/2024-07-02_14_22_39-2884532041560971713;graphView=0?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22))

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Anna Geiduschek <anna@recidiviz.org>
GitOrigin-RevId: 4618864496c03a9734374b3748ec722187b4d674