Adds insights caseload category sessions and assignment view (Recidiviz/recidiviz-data#30997)
## Description of the change

This PR adds the assignment view
(`supervision_insights_caseload_category_metrics_person_assignment_sessions`)
required as a precursor for outliers product views to spin up metrics
disaggregated by caseload category.

Changes:
1) Adds a new insights_caseload_category_sessions view, which defines
new enums for InsightsCaseloadCategory and InsightsCaseloadCategoryType
(indicates the type of categorization logic). For now, the enums are
only defined for the category type that will be used in Idaho (a binary
between SEX_OFFENSE vs. others)
2) Defines a new unit of analysis for insights caseload categories, and
adds this type to the appropriate configs in the
aggregated_metrics_view_collector file so that we are able to spit out
the assignment sessions view but don't spin up the default aggregated
metrics views (since we want to only materialize a small subset of
metrics in a product view)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#30812
Closes Recidiviz/recidiviz-data#30813

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: 4c2815d0e481a5ee7360066a1324706a68abbdec