Add new metric for distinct registered workflows users (Recidiviz/recidiviz-data#30990)
## Description of the change

This is the second of two PR's that unblock a metric that captures the
total number of distinct registered workflows users (See Recidiviz/recidiviz-data#30896 for
first PR). This is also dependent on
https://github.com/Recidiviz/recidiviz-data/pull/31092, which defines a
new workflows user Span for registration sessions. This PR does the
following:
1) Defines a new metric class type called SpanDistinctUnitCount, which
(similar to EventDistinctUnitCount) returns a distinct count across all
the units of observation represented in a subset of spans.
3) Configures a new workflows explore metric called
`distinct_registered_users`, which is a SpanDistinctUnitCount metric
that will count the number of workflows users having an open
registration session overlapping with a given analysis period.

After running the lookml autogen update script, this PR will add this
new metric to the menu of options.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: 867767515531848ea6fcc6787ea47a3f3190080b