[US_AR] Address EXTERNAL_UNKNOWN supervision districts (Recidiviz/recidiviz-data#31247)
## Description of the change

Some supervision districts are showing up as EXTERNAL_UNKNOWN because
they represent a supervision area name that lacks an associated ID: West
Memphis (which is no longer its own area, but rather part of area 4) and
C.O. Interstate Compact. This change manually remaps the "West Memphis"
name to "Jonesboro", which is the current name for Area 4, and adds the
"ISC" ID for C.O. Interstate Compact. With this change, every
supervision area name that appears in the data will have an associated
supervision_district_id.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: be821f752ee34874b96399079dbabdb0e6bf386d