[US_MO] Sort sentence status sequence_nums numerically instead of alphabetically (Recidiviz/recidiviz-data#31285)
## Description of the change

While validating Recidiviz/recidiviz-data#30413, I found a bug in how we order sentence statuses
- we were sorting them by `sentence_status_external_id` which is a
string with the sequence number in it. This means that a status like
"X-X-X-10" would get sorted before "X-X-X-9". I'm introducing this fix
first so that it is easier to validate the actual change that will use
the new sentence statuses.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#30199

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b8bc8d313d6e8ed3ef30d024bdf4eb1d536c29ba